+++ docs/cmd/ach_contest_recent.md
@@ -1,0 +1,29 @@
+## ach contest recent
+
+show recent contests
+
+### Synopsis
+
+show recent contests.
+
+```
+ach contest recent [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for recent
+```
+
+### Options inherited from parent commands
+
+```
+      --config string        config file (default "$HOME/.ach/config.yaml")
+      --task-config string   task config file (default "./achTaskConfig.yaml")
+```
+
+### SEE ALSO
+
+* [ach contest](ach_contest.md)	 - manipulates an AtCoder contest
+
